// src/libs/core-api/api/model/channel-information-response.ts
/* tslint:disable */

/* eslint-disable */

/**
 * BlockInfrastructure-Core
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
// May contain unused imports in some cases
// @ts-ignore
import { ChannelUserInformationProjection } from './channel-user-information-projection';

/**
 *
 * @export
 * @interface ChannelInformationResponse
 */
export interface ChannelInformationResponse {
  /**
   *
   * @type {string}
   * @memberof ChannelInformationResponse
   */
  channelId: string;
  /**
   *
   * @type {string}
   * @memberof ChannelInformationResponse
   */
  name: string;
  /**
   *
   * @type {string}
   * @memberof ChannelInformationResponse
   */
  description: string;
  /**
   *
   * @type {string}
   * @memberof ChannelInformationResponse
   */
  profileImageUrl?: string | null;
  /**
   *
   * @type {Array<ChannelUserInformationProjection>}
   * @memberof ChannelInformationResponse
   */
  channelUserInformationList: Array<ChannelUserInformationProjection>;
}